detection/requirements.txt
# Detection Service Dependencies
py-clob-client>=0.8.0
websockets>=12.0
web3>=6.15.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
prometheus-client>=0.19.0
structlog>=24.1.0
orjson>=3.9.0
asyncio-throttle>=1.0.2
